requirements.txt
Adafruit-Blinka==8.56.0
adafruit-circuitpython-bmp280==3.3.6
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-register==1.10.2
adafruit-circuitpython-requests==4.1.10
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.77.0
Adafruit-PureIO==1.1.11
annotated-types==0.7.0
anyio==4.9.0
binho-host-adapter==0.1.6
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.14.0
httptools==0.6.4
idna==3.10
jinja2==3.1.6
MarkupSafe==3.0.2
psutil==7.0.0
pydantic==2.11.3
pydantic-core==2.33.1
pyftdi==0.56.0
pyserial==3.5
python-dotenv==1.1.0
pyusb==1.3.1
PyYAML==6.0.2
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
sniffio==1.3.1
starlette==0.46.1
sysv-ipc==1.1.0
toml==0.10.2
typing-extensions==4.13.1
typing-inspection==0.4.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
